+++ inventory-service/src/main/java/org/mddarr/inventory/port/ProductApiInterface.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 public interface ProductApiInterface {
 
     @RequestMapping(value = "add", produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.POST)
-    void addProduct(@RequestParam("name") String name, @RequestParam("brand") String brand, @RequestParam("price") Long price);
+    void addProduct(@RequestParam("name") String name, @RequestParam("brand") String brand, @RequestParam("price") Double price);
 
     @RequestMapping(value = "delete", produces = MediaType.APPLICATION_JSON_VALUE, method = RequestMethod.POST)
     void deleteProduct(String id);
